+++ android/versioned-abis/expoview-abi44_0_0/src/main/java/abi44_0_0/expo/modules/calendar/EventNotSavedException.java
@@ -1,4 +1,0 @@
-package abi44_0_0.expo.modules.calendar;
-
-public class EventNotSavedException extends Exception {
-}